pyproject.toml
[project]
name = "plugram"
version = "0.1.0"
description = "Modular Telegram userbot"
requires-python = ">=3.11"
dependencies = [
    "telethon>=1.36",
    "openai>=1.40",
    "python-dotenv>=1.0",
    "tomli-w>=1.0",
]

[project.scripts]
plugram = "plugram.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["plugram"]
